Refactor tool view
It prevent ui issues and different behaviour on browsers
