+++ frontend/package.json
@@ -20,7 +20,6 @@
     "@mapbox/mapbox-gl-language": "^0.10.1",
     "@websanova/vue-auth": "^2.21.14-beta",
     "axios": "^0.19.0",
-    "bowser": "^2.6.1",
     "buefy": "^0.7.3",
     "bulma-steps": "https://git@github.com/FranckKe/bulma-steps.git",
     "lodash-es": "^4.17.15",
