+++ frontend/src/components/MenuNav.vue
@@ -1,14 +1,6 @@
 <template>
-  <nav
-    class="navbar is-fixed-top"
-    role="navigation"
-    aria-label="main navigation"
-  >
-    <b-loading
-      :is-full-page="false"
-      :active.sync="this.loading"
-      :can-cancel="false"
-    ></b-loading>
+  <nav class="navbar is-fixed-top" role="navigation" aria-label="main navigation">
+    <b-loading :is-full-page="false" :active.sync="this.loading" :can-cancel="false"></b-loading>
     <div class="navbar-brand">
       <div class="navbar-start">
         <a
@@ -26,16 +18,11 @@
     </div>
     <div id="flexibleMenu" class="navbar-menu">
       <router-link to="/" class="navbar-item">{{ $t('home') }}</router-link>
-      <router-link
-        :to="`/${$t('tracers').toLowerCase()}`"
-        class="navbar-item"
-        >{{ $t('tracers') }}</router-link
-      >
+      <router-link :to="`/${$t('tracers').toLowerCase()}`" class="navbar-item">{{ $t('tracers') }}</router-link>
       <router-link
         :to="`/${$t('leaderboard').toLowerCase()}`"
         class="navbar-item"
-        >{{ $t('leaderboard') }}</router-link
-      >
+      >{{ $t('leaderboard') }}</router-link>
 
       <div
         class="navbar-item has-dropdown is-hoverable"
@@ -51,8 +38,7 @@
             :to="
               `/pages/${slugify($t(category)).toLowerCase()}/${pageName.slug}`
             "
-            >{{ pageName.raw }}</router-link
-          >
+          >{{ pageName.raw }}</router-link>
         </div>
       </div>
       <lang-switcher></lang-switcher>
@@ -64,30 +50,20 @@
           </button>
 
           <b-dropdown-item>
-            <router-link
-              :to="'/me'"
-              class="has-text-dark is-size-6 nav-dropdown-link"
-            >
+            <router-link :to="'/me'" class="has-text-dark is-size-6 nav-dropdown-link">
               <font-awesome-icon icon="user" />
               <p>{{ $t('my_account') }}</p>
             </router-link>
           </b-dropdown-item>
           <b-dropdown-item class="has-text-dark is-size-6">
-            <router-link
-              :to="'/me/reports'"
-              class="has-text-dark is-size-6 nav-dropdown-link"
-            >
+            <router-link :to="'/me/reports'" class="has-text-dark is-size-6 nav-dropdown-link">
               <font-awesome-icon icon="map-marker-alt" />
               <p>{{ $t('my_reports') }}</p>
             </router-link>
           </b-dropdown-item>
           <hr class="dropdown-divider" />
           <b-dropdown-item class="has-text-danger is-size-6">
-            <a
-              href="#"
-              class="has-text-danger is-size-6 nav-dropdown-link"
-              v-on:click="logout()"
-            >
+            <a href="#" class="has-text-danger is-size-6 nav-dropdown-link" v-on:click="logout()">
               <font-awesome-icon icon="sign-out-alt" />
               <p>{{ $t('logout') }}</p>
             </a>
@@ -96,11 +72,7 @@
         <router-link v-if="!$auth.check()" :to="'/login'" class="button">
           <p>{{ $t('login') }}</p>
         </router-link>
-        <router-link
-          v-if="!$auth.check()"
-          :to="'/register'"
-          class="button is-success"
-        >
+        <router-link v-if="!$auth.check()" :to="'/register'" class="button is-success">
           <p>{{ $t('register') }}</p>
         </router-link>
       </div>
@@ -133,8 +105,6 @@ export default {
     navbarBurger.addEventListener('click', () => {
       navbarBurger.classList.toggle('is-active')
       $target.classList.toggle('is-active')
-
-      if (this.getActiveTool() !== '') this.closeToolbar()
     })
   },
   computed: {
@@ -152,8 +122,6 @@ export default {
   },
   methods: {
     ...pagesModule.mapActions(['loadPages']),
-    ...toolBarModule.mapActions(['closeToolbar']),
-    ...toolBarModule.mapGetters(['getActiveTool']),
     slugify: slugify,
     logout: function() {
       this.$auth.logout({
